+++ NEWS.md
@@ -1,5 +1,6 @@
 Version 0.6.4
  - fix: code completion popup not show after '->' inputted
+ - fix: font styles in the color scheme settings not in effect
 
 Version 0.6.3
  - fix: should use c++ syntax to check ".h" files